# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UpFlow - Regulação de Leitos API",
        "description": "API REST para gerenciamento e controle de leitos hospitalares desenvolvida em Laravel.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Ambiente Local"
        }
    ],
    "paths": {
        "/api/internacoes": {
            "get": {
                "tags": [
                    "Internacoes"
                ],
                "summary": "Lista todas as internações",
                "operationId": "1ff8982f940dbc97a1595520395e2888",
                "responses": {
                    "200": {
                        "description": "Lista de internações retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Internacoes"
                ],
                "summary": "Registra uma nova internação",
                "operationId": "e4be0b3d890290ac19e656acb7bfc757",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "paciente_id",
                                    "leito_id",
                                    "data_internacao"
                                ],
                                "properties": {
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "leito_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "data_internacao": {
                                        "type": "string",
                                        "example": "2026-06-17 22:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Internação registrada com sucesso"
                    },
                    "422": {
                        "description": "Regra de negócio ou validação não atendida"
                    }
                }
            }
        },
        "/api/internacoes/{internacao}": {
            "get": {
                "tags": [
                    "Internacoes"
                ],
                "summary": "Exibe uma internação específica",
                "operationId": "67b52e072c4949891170ed28816f4880",
                "parameters": [
                    {
                        "name": "internacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Internação retornada com sucesso"
                    },
                    "404": {
                        "description": "Internação não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Internacoes"
                ],
                "summary": "Atualiza uma internação",
                "operationId": "7cea956622810b26ba577a99f975424f",
                "parameters": [
                    {
                        "name": "internacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "paciente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "leito_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "data_internacao": {
                                        "type": "string",
                                        "example": "2026-06-17 22:00:00"
                                    },
                                    "data_alta": {
                                        "type": "string",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "INTERNADO"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Internação atualizada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Internacoes"
                ],
                "summary": "Remove uma internação",
                "operationId": "7e2868d337c15ca5937e1e762c364466",
                "parameters": [
                    {
                        "name": "internacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Internação removida com sucesso"
                    },
                    "404": {
                        "description": "Internação não encontrada"
                    }
                }
            }
        },
        "/api/internacoes/{internacao}/alta": {
            "patch": {
                "tags": [
                    "Internacoes"
                ],
                "summary": "Registra alta de uma internação",
                "operationId": "82400962567a2fdd2ca417614ccf5fbc",
                "parameters": [
                    {
                        "name": "internacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alta registrada com sucesso"
                    },
                    "422": {
                        "description": "Internação já possui alta registrada"
                    }
                }
            }
        },
        "/api/internacoes/{internacao}/transferir": {
            "patch": {
                "tags": [
                    "Internacoes"
                ],
                "summary": "Transfere um paciente para outro leito",
                "description": "Transfere uma internação ativa para outro leito disponível.",
                "operationId": "9496c50ec822a241f8ddf944387a8982",
                "parameters": [
                    {
                        "name": "internacao",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "leito_id"
                                ],
                                "properties": {
                                    "leito_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente transferido com sucesso"
                    },
                    "422": {
                        "description": "Regra de negócio ou validação não atendida"
                    }
                }
            }
        },
...
    },
...
}